Add AtLevelTest, move related testcases to AtLevelTest
